Refactor Ollama payload helper
